change uint32_t to uint64_t

+++ src/ADMM.cpp
@@ -7,9 +7,9 @@ ADMM::ADMM (const arma::mat &X,
             const arma::vec &betaWS,
             const arma::vec &zWS,
             const arma::vec &uWS,
-            const uint32_t KLB,
-            const uint32_t maxIter,
-            const uint32_t threadNum) {
+            const uint64_t KLB,
+            const uint64_t maxIter,
+            const uint64_t threadNum) {
     reset(X, y, o, w, betaWS, zWS, uWS, KLB, maxIter, threadNum);
 }
 
@@ -20,10 +20,10 @@ void ADMM::reset(const arma::mat &X,
                  const arma::vec &betaWS,
                  const arma::vec &zWS,
                  const arma::vec &uWS,
-                 const uint32_t KLB,
-                 const uint32_t maxIter,
-                 const uint32_t threadNum) {
-    uint32_t n = X.n_rows, p = X.n_cols;
+                 const uint64_t KLB,
+                 const uint64_t maxIter,
+                 const uint64_t threadNum) {
+    uint64_t n = X.n_rows, p = X.n_cols;
     this->X = X;
     this->y = y;
     setVec(this->o, o, n);
@@ -54,7 +54,7 @@ void ADMM::clear() {
 }
 
 arma::vec ADMM::fit(const std::string method) {
-    uint32_t k = 1, p = X.n_cols, n = X.n_rows;
+    uint64_t k = 1, p = X.n_cols, n = X.n_rows;
     std::vector<ADMMSolver *> solvers(threadNum);
     std::string lowerMethod(method);
     sumUBeta = arma::mat(p, threadNum, arma::fill::zeros);
@@ -99,14 +99,14 @@ void ADMM::setWeight(const arma::vec &weight) {
     }
     else if (weight.n_elem == X.n_cols - 1) {
         this-> w = arma::vec(X.n_cols, arma::fill::zeros);
-        for (uint32_t i = 0; i < weight.n_elem; i++) {
+        for (uint64_t i = 0; i < weight.n_elem; i++) {
             this->w[i + 1] = weight[i];
         }
     }
 }
 
 void ADMM::setWarmStartPara(const arma::vec &zWS, const arma::vec &uWS, const arma::vec &w) {
-    uint32_t p = X.n_cols;
+    uint64_t p = X.n_cols;
     setVec(this->betaWS, betaWS, p);
     setVec(this->zWS, zWS, p);
     setVec(this->uWS, uWS, p);
@@ -116,21 +116,21 @@ void ADMM::setInitialBeta(const arma::vec &beta) {
     setVec(this->betaWS, betaWS, X.n_cols);
 }
 
-void ADMM::setKLB(const int KLB) {
+void ADMM::setKLB(const uint64_t KLB) {
     if (KLB < 0) {
         Rcpp::stop("The KLB should not be less than 0.");
     }
     this->KLB = KLB;
 }
 
-void ADMM::setMaxIterator(const int maxIter) {
+void ADMM::setMaxIterator(const uint64_t maxIter) {
     if (maxIter < 0) {
         Rcpp::stop("The maxIter should not be less than 0.");
     }
     this->maxIter = maxIter;
 }
 
-void ADMM::setThreadNumber(const int number) {
+void ADMM::setThreadNumber(const uint64_t number) {
     if (number <= 0) {
         Rcpp::stop("The thread number should not be less than 1.");
     }
@@ -141,25 +141,25 @@ void ADMM::setThreadNumber(const int number) {
 
 template <typename T>
 void ADMM::initializeSolver(std::vector<ADMMSolver *> &solvers) {
-    uint32_t n = X.n_rows;
+    uint64_t n = X.n_rows;
     double interval = n / static_cast<double>(threadNum);
-    for (uint32_t i = 0; i < threadNum; i++) {
-        uint32_t col0 = i * interval, col1 = (i + 1) * interval - 1;
+    for (uint64_t i = 0; i < threadNum; i++) {
+        uint64_t col0 = i * interval, col1 = (i + 1) * interval - 1;
         arma::mat subX = X.rows(col0, col1);
         solvers[i] = new T(X.rows(col0, col1), y.subvec(col0, col1), o, betaWS, uWS);
     }
 }
 
 void ADMM::deleteSolver(std::vector<ADMMSolver *> &solvers) {
-    for (uint32_t i = 0; i < threadNum; i++) {
+    for (uint64_t i = 0; i < threadNum; i++) {
         delete solvers[i];
     }
 }
 
 arma::vec ADMM::updateUBeta(std::vector<ADMMSolver *> &solvers) {
-    uint32_t p = X.n_cols;
+    uint64_t p = X.n_cols;
     #pragma omp parallel for
-    for (uint32_t i = 0; i < threadNum; i++) {
+    for (uint64_t i = 0; i < threadNum; i++) {
          sumUBeta.col(i) = solvers[i]->solve(z);
     }
     return arma::sum(sumUBeta, 1) / threadNum;
@@ -170,8 +170,8 @@ void ADMM::updateZ() {
 }
 
 void ADMM::softThreashold(const arma::vec &sum, arma::vec &value) {
-    uint32_t p = X.n_cols;
-    for (uint32_t i = 0; i < p; i++) {
+    uint64_t p = X.n_cols;
+    for (uint64_t i = 0; i < p; i++) {
         if (sum[i] > w[i]) {
             value[i] = sum[i] - w[i];
         }
@@ -196,7 +196,7 @@ bool ADMM::stopCriteria() {
     return supportIter >= KLB;
 }
 
-void ADMM::setVec(arma::vec &target, const arma::vec &source, const uint32_t num) {
+void ADMM::setVec(arma::vec &target, const arma::vec &source, const uint64_t num) {
     if (source.empty()) {
         target = arma::vec(num, arma::fill::zeros);
     }
